refactor(inspector): split Swift type panel into a TabViewController
Reshape InspectorSwiftTypeViewController into an empty TabViewController
subclass mirroring SidebarRootTabViewController, with two child VCs
selected via the tab bar:

- Hierarchy: reuses InspectorClassViewController/InspectorClassViewModel
  so Swift class hierarchy rendering shares its ObjC counterpart.
- Specialization: new InspectorSwiftSpecializationViewController and
  InspectorSwiftSpecializationViewModel host the specialized-children
  list and the "+ Add Specialization" entry point.

InspectorCoordinator now picks single-child or two-tab layout from
isClass / (isGeneric && !isSpecialized), so a non-class generic value
type drops straight into the specialization VC instead of synthesizing
a hidden segment. The old InspectorSwiftTypeViewModel and its
SegmentVisibility plumbing are removed.
